Rename `GroovyConfigPropertySourceLoader` `YamlPropertySourceLoader` package from `org.grails.spring.boot.env` to `org.grails.config.loader`
- Rename package `GroovyConfigPropertySourceLoaderSpec` `YamlPropertySourceLoaderSpec`
- Update `grace-plugin-core/src/main/resources/META-INF/spring.factories`
