+++ grace-bootstrap/src/test/groovy/org/grails/config/PropertySourceConfigSpec.groovy
@@ -2,7 +2,7 @@ package grails.config
 
 import org.grails.config.NavigableMap
 import org.grails.config.PropertySourcesConfig
-import org.grails.spring.boot.env.YamlPropertySourceLoader
+import org.grails.config.loader.YamlPropertySourceLoader
 import org.springframework.core.io.FileSystemResource
 import org.springframework.core.io.Resource
 import spock.lang.Specification
