+++ grace-plugin-core/src/main/resources/META-INF/spring.factories
@@ -1,4 +1,4 @@
 # PropertySource Loaders
 org.springframework.boot.env.PropertySourceLoader=\
-org.grails.spring.boot.env.GroovyConfigPropertySourceLoader,\
-org.grails.spring.boot.env.YamlPropertySourceLoader
+org.grails.config.loader.GroovyConfigPropertySourceLoader,\
+org.grails.config.loader.YamlPropertySourceLoader
